Fixed forecaster id feed filtering
