+++ tests/unit/test_auth/test_services.py
@@ -4,7 +4,6 @@ from django.contrib.auth.tokens import default_token_generator
 from rest_framework.exceptions import ValidationError
 
 from authentication.services import check_and_activate_user
-from tests.unit.fixtures import *  # noqa
 from tests.unit.test_users.factories import factory_user
