+++ tests/unit/test_comments/test_serializers.py
@@ -1,7 +1,6 @@
 import pytest  # noqa
 
 from comments.serializers import serialize_key_factors_many
-from tests.unit.fixtures import *  # noqa
 from tests.unit.test_comments.factories import factory_comment, factory_key_factor
 from tests.unit.test_posts.factories import factory_post
 from tests.unit.test_questions.conftest import *  # noqa
