+++ tests/unit/test_comments/test_services.py
@@ -5,7 +5,6 @@ from comments.services.key_factors import key_factor_vote
 from comments.services.notifications import notify_mentioned_users
 from posts.models import Post, PostUserSnapshot
 from projects.permissions import ObjectPermission
-from tests.unit.fixtures import *  # noqa
 from tests.unit.test_comments.factories import factory_comment, factory_key_factor
 from tests.unit.test_posts.factories import factory_post
 from tests.unit.test_projects.factories import factory_project
