+++ tests/unit/test_projects/test_services.py
@@ -2,7 +2,6 @@ from notifications.models import Notification
 from posts.models import Post
 from projects.permissions import ObjectPermission
 from projects.services.common import notify_project_subscriptions_post_open
-from tests.unit.fixtures import *  # noqa
 from tests.unit.test_posts.factories import factory_post
 from tests.unit.test_projects.factories import factory_project
 from tests.unit.test_users.factories import factory_user
