+++ tests/unit/test_utils/test_middlewares.py
@@ -1,7 +1,5 @@
 from django.conf import settings
 
-from tests.unit.fixtures import *  # noqa
-
 
 class TestAuthenticationRequiredMiddleware:
     def test_anon(self, anon_client):
